// src/app/api/models/permission-response-dto.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface PermissionResponseDto {
  action: string;
  createdAt: string;
  deletedAt: string | null;
  description: string | null;
  id: string;
  module: string;
  name: string;
  subject: string;
  updatedAt: string;
}